requirements.txt
gradio 
google-genai 
python-dotenv 
pypdf2
faiss-cpu 
numpy
